refactor(api) migrate UserNotFoundError to `src/shared`
