+++ api/lib/domain/usecases/authenticate-external-user.js
@@ -5,11 +5,8 @@ import {
   UserShouldChangePasswordError,
 } from '../../../src/identity-access-management/domain/errors.js';
 import { AuthenticationMethod } from '../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
-import {
-  UnexpectedUserAccountError,
-  UserAlreadyExistsWithAuthenticationMethodError,
-  UserNotFoundError,
-} from '../errors.js';
+import { UserNotFoundError } from '../../../src/shared/domain/errors.js';
+import { UnexpectedUserAccountError, UserAlreadyExistsWithAuthenticationMethodError } from '../errors.js';
 
 async function authenticateExternalUser({
   username,
