+++ api/src/identity-access-management/domain/services/sco-account-recovery.service.js
@@ -4,9 +4,9 @@ import {
   AccountRecoveryDemandExpired,
   MultipleOrganizationLearnersWithDifferentNationalStudentIdError,
   UserHasAlreadyLeftSCO,
-  UserNotFoundError,
 } from '../../../../lib/domain/errors.js';
 import { config } from '../../../shared/config.js';
+import { UserNotFoundError } from '../../../shared/domain/errors.js';
 
 const { uniqBy } = lodash;
 const { features } = config;
