+++ api/tests/certification/session-management/integration/infrastructure/repositories/certification-officer-repository_test.js
@@ -1,6 +1,6 @@
-import { UserNotFoundError } from '../../../../../../lib/domain/errors.js';
 import { CertificationOfficer } from '../../../../../../src/certification/session-management/domain/models/CertificationOfficer.js';
 import * as certificationOfficerRepository from '../../../../../../src/certification/session-management/infrastructure/repositories/certification-officer-repository.js';
+import { UserNotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | CertificationOfficer', function () {
