+++ api/tests/identity-access-management/integration/infrastructure/repositories/user.repository.test.js
@@ -1,11 +1,6 @@
 import lodash from 'lodash';
 const { each, map, times, pick } = lodash;
-import {
-  AlreadyExistingEntityError,
-  AlreadyRegisteredUsernameError,
-  NotFoundError,
-  UserNotFoundError,
-} from '../../../../../lib/domain/errors.js';
+import { AlreadyExistingEntityError, AlreadyRegisteredUsernameError } from '../../../../../lib/domain/errors.js';
 import { CertificationCenter } from '../../../../../lib/domain/models/CertificationCenter.js';
 import { CertificationCenterMembership } from '../../../../../lib/domain/models/CertificationCenterMembership.js';
 import { Membership } from '../../../../../lib/domain/models/Membership.js';
@@ -17,7 +12,7 @@ import * as OidcIdentityProviders from '../../../../../src/identity-access-manag
 import { User } from '../../../../../src/identity-access-management/domain/models/User.js';
 import * as userRepository from '../../../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
 import { Organization } from '../../../../../src/organizational-entities/domain/models/Organization.js';
-import { AlreadyRegisteredEmailError } from '../../../../../src/shared/domain/errors.js';
+import { AlreadyRegisteredEmailError, UserNotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../test-helper.js';
 
 const expectedUserDetailsForAdminAttributes = [
@@ -608,7 +603,7 @@ describe('Integration | Identity Access Management | Infrastructure | Repository
         const result = await catchErr(userRepository.getByEmail)(emailThatDoesNotExist);
 
         // then
-        expect(result).to.be.instanceOf(NotFoundError);
+        expect(result).to.be.instanceOf(UserNotFoundError);
       });
 
       it('should return the user with email case insensitive', async function () {
@@ -1715,7 +1710,7 @@ describe('Integration | Identity Access Management | Infrastructure | Repository
 
     it('should throw an error when the user does not exist by email', async function () {
       const err = await catchErr(userRepository.isUserExistingByEmail)('none');
-      expect(err).to.be.instanceOf(NotFoundError);
+      expect(err).to.be.instanceOf(UserNotFoundError);
     });
   });
