+++ api/tests/identity-access-management/unit/domain/services/sco-account-recovery.service.test.js
@@ -4,11 +4,10 @@ import {
   AccountRecoveryDemandExpired,
   MultipleOrganizationLearnersWithDifferentNationalStudentIdError,
   UserHasAlreadyLeftSCO,
-  UserNotFoundError,
 } from '../../../../../lib/domain/errors.js';
 import { scoAccountRecoveryService } from '../../../../../src/identity-access-management/domain/services/sco-account-recovery.service.js';
 import { config } from '../../../../../src/shared/config.js';
-import { AlreadyRegisteredEmailError } from '../../../../../src/shared/domain/errors.js';
+import { AlreadyRegisteredEmailError, UserNotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
 const { features } = config;
