+++ api/tests/integration/application/organization-invitations/organization-invitation-controller_test.js
@@ -1,6 +1,7 @@
 import * as moduleUnderTest from '../../../../lib/application/organization-invitations/index.js';
-import { AlreadyExistingInvitationError, NotFoundError, UserNotFoundError } from '../../../../lib/domain/errors.js';
+import { AlreadyExistingInvitationError, NotFoundError } from '../../../../lib/domain/errors.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { UserNotFoundError } from '../../../../src/shared/domain/errors.js';
 import { serializer as scoOrganizationInvitationSerializer } from '../../../../src/team/infrastructure/serializers/jsonapi/sco-organization-invitation.serializer.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
