+++ api/tests/integration/infrastructure/knex-database-connection_test.js
@@ -1,6 +1,6 @@
 import { emptyAllTables, knex } from '../../../db/knex-database-connection.js';
-import { UserNotFoundError } from '../../../lib/domain/errors.js';
 import * as userRepository from '../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
+import { UserNotFoundError } from '../../../src/shared/domain/errors.js';
 import { databaseBuilder, expect } from '../../test-helper.js';
 
 describe('Integration | Infrastructure | knex-database-connection', function () {
