+++ api/tests/unit/domain/errors_test.js
@@ -2,6 +2,7 @@ import * as errors from '../../../lib/domain/errors.js';
 import { NotEnoughDaysPassedBeforeResetCampaignParticipationError } from '../../../lib/domain/errors.js';
 import { AdminMemberError } from '../../../src/authorization/domain/errors.js';
 import { UnableToAttachChildOrganizationToParentOrganizationError } from '../../../src/organizational-entities/domain/errors.js';
+import { UserNotFoundError } from '../../../src/shared/domain/errors.js';
 import {
   CsvImportError,
   EntityValidationError,
@@ -75,7 +76,7 @@ describe('Unit | Domain | Errors', function () {
 
   describe('#UserNotFoundError', function () {
     it('should export a UserNotFoundError', function () {
-      expect(errors.UserNotFoundError).to.exist;
+      expect(UserNotFoundError).to.exist;
     });
 
     it('should have a getErrorMessage method', function () {
@@ -87,7 +88,7 @@ describe('Unit | Domain | Errors', function () {
       };
 
       // then
-      const userNotFoundError = new errors.UserNotFoundError();
+      const userNotFoundError = new UserNotFoundError();
       expect(userNotFoundError.getErrorMessage).to.be.a('function');
       expect(userNotFoundError.getErrorMessage()).to.eql(expectedErrorMessage);
     });
