+++ api/tests/unit/domain/usecases/authenticate-external-user_test.js
@@ -1,7 +1,6 @@
 import {
   UnexpectedUserAccountError,
   UserAlreadyExistsWithAuthenticationMethodError,
-  UserNotFoundError,
 } from '../../../../lib/domain/errors.js';
 import { authenticateExternalUser } from '../../../../lib/domain/usecases/authenticate-external-user.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
@@ -11,6 +10,7 @@ import {
   UserShouldChangePasswordError,
 } from '../../../../src/identity-access-management/domain/errors.js';
 import { AuthenticationMethod } from '../../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
+import { UserNotFoundError } from '../../../../src/shared/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | Application | UseCase | authenticate-external-user', function () {
