+++ api/tests/unit/domain/usecases/reassign-authentication-method-to-another-user_test.js
@@ -1,6 +1,7 @@
-import { AuthenticationMethodAlreadyExistsError, UserNotFoundError } from '../../../../lib/domain/errors.js';
+import { AuthenticationMethodAlreadyExistsError } from '../../../../lib/domain/errors.js';
 import { reassignAuthenticationMethodToAnotherUser } from '../../../../lib/domain/usecases/reassign-authentication-method-to-another-user.js';
 import * as OidcIdentityProviders from '../../../../src/identity-access-management/domain/constants/oidc-identity-providers.js';
+import { UserNotFoundError } from '../../../../src/shared/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | reassign-authentication-method-to-another-user', function () {
