+++ api/tests/unit/domain/usecases/update-organization-learner-dependent-user-password_test.js
@@ -1,5 +1,5 @@
-import { UserNotFoundError } from '../../../../lib/domain/errors.js';
 import { updateOrganizationLearnerDependentUserPassword } from '../../../../lib/domain/usecases/update-organization-learner-dependent-user-password.js';
+import { UserNotFoundError } from '../../../../src/shared/domain/errors.js';
 import { UserNotAuthorizedToUpdatePasswordError } from '../../../../src/shared/domain/errors.js';
 import { catchErr, expect, sinon } from '../../../test-helper.js';
